+++ src/thread_pool.cpp
@@ -1,6 +1,7 @@
 // Copyright (c) Quantum Brilliance Pty Ltd
 #include <qristal/core/thread_pool.hpp>
 
+#include <algorithm>
 #include <iostream>
 
 namespace qristal